@@ -1,20 +1,24 @@
-[tool.poetry]
+[project]
 name = "pypdfmerger"
 version = "1.2.0"
-description = ""
-authors = ["ludgerradke <radke@intervall.io>"]
+description = "A PyQt6 Application for Merging PDF Files"
+authors = [{ name = "ludgerradke", email = "radke@intervall.io" }]
 readme = "README.md"
+requires-python = ">=3.12"
+dependencies = [
+    "pypdf2>=3.0.1",
+    "pdf2image>=1.16.2",
+    "pymupdf>=1.22.3",
+    "pyqt6>=6.5.1",
+    "screeninfo>=0.8.1",
+]
 
-[tool.poetry.dependencies]
-python =  ">=3.10,<3.12"
-pypdf2 = "^3.0.1"
-pdf2image = "^1.16.2"
-pyinstaller = "^5.11.0"
-pymupdf = "^1.22.3"
-pyqt6 = "^6.5.1"
-screeninfo = "^0.8.1"
-
+[project.optional-dependencies]
+dev = [
+    "pyinstaller>=6.0.0",
+    "ty>=0.0.1a7",
+]
 
 [build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
+requires = ["hatchling"]
+build-backend = "hatchling.build"
